refactor: removed type decoders from project
- removed type-decoders from the project.
- type decoders to be used from https://github.com/sinha-sahil/type-decoder or  https://www.npmjs.com/package/type-decoder
