+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "typesafe-api-call",
-  "version": "3.0.0",
+  "version": "4.0.0",
   "description": "A TS/JS API caller util for making typesafe API calls",
   "main": "index.ts",
   "type": "module",
